Lift ConfirmationDialog out into its own component
